+++ association/src/test/java/com/association/controller/logincontrollertest.java
@@ -16,11 +16,11 @@ public class logincontrollertest {
     @Autowired
     private UserService login;
     @GetMapping("testmapperlogin1/{id}/{pw}")
-    public User getuserbylogin(@PathVariable("id") double id, @PathVariable("pw") String pw){
+    public User getuserbylogin(@PathVariable("id") long id, @PathVariable("pw") String pw){
         return login.getuserbylogin(id,pw);
 }
     @GetMapping("testmapperlogin2/{id}")
-    public User getuserbyid(@PathVariable("id") double id){
+    public User getuserbyid(@PathVariable("id") long id){
         return login.getUserById(id);
     }
 }
